refactor: update cdp version; remove unused cdp_simple and test files
