+++ requirements.txt
@@ -11,7 +11,6 @@ chromadb>=0.4.22
 tiktoken>=0.5.2
 web3>=7.7.0
 eth-abi>=5.2.0
-cdp-langchain>=0.0.13
 pydantic>=2.0.0
 pypdf>=5.2.0
 chromadb>=0.6.3
@@ -21,3 +20,5 @@ websockets>=13.0
 aiohttp>=3.9.0
 aiohttp-middlewares>=2.0.0
 supabase>=2.13.0
+coinbase-agentkit>=0.1.0
+coinbase-agentkit-langchain>=0.1.0
