+++ src/utils/__init__.py
@@ -1,10 +1,9 @@
 from .telegram import send_telegram_message_async
-from .cdp import get_user_shares_tool, get_reallocation_tool
 from .memory import add_long_term_memory, get_long_term_memory
 from .reasoning import market_analysis
 from .market import format_market_history, get_all_market_history
-
+from .action_provider import morpho_action_provider
 __all__ = [
-  'send_telegram_message_async', 'get_reallocation_tool', 'get_user_shares_tool', 
+  'send_telegram_message_async', 
   'add_long_term_memory', 'get_long_term_memory', 'market_analysis', 'format_market_history', 'get_all_market_history'
   ]
